requirements.txt
discord.py
django
mysqlclient
bs4
uvloop
tzdata
lxml
